fix: replace Node.js crypto APIs with Web Crypto, extract WelcomeInput, add tests
- Replace timingSafeEqual/Buffer with constant-time Uint8Array XOR compare
  (emailChange.ts mixes queries+mutations, cannot use "use node")
- Extract WelcomeInput into its own file (page.tsx was 325 lines, now 172)
- Add sanitizeForDiscord unit tests (contact.test.ts)
- Fix stale test description: SHA-256 -> HMAC-SHA-256
